Hypersurface minor tweaks
